tests/wpt/tests/fenced-frame/embedder-no-coep.https.html

<!DOCTYPE html>
<title>Test COEP properties set for a Fenced Frame Tree</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>
<script src="resources/embedder-policy.js"></script>

<body>
<script>
promise_test(async () => {
  const uuid = token();
  const frame = setupTest("coep:require-corp", uuid);
  const result = await nextValueFromServer(uuid);
  assert_equals(result, "PASS", "embedded page has been loaded.");
}, "Create fencedframe with COEP:require-corp");

promise_test(async () => {
  const uuid = token();
  const frame = setupTest("no coep", uuid);
  const result = await nextValueFromServer(uuid);
  assert_equals(result, "PASS",
                "page without COEP should be able to load page with COEP");
}, "Create fencedframe without COEP header");

</script>
</body>
